// src/app/shared/models/user.ts
export class User {
  id?: number;
  username: string;
  firstName: string;
  middleName: string;
  lastName: string;
  ssn?: string;
  email: string;
  phone: string;
  role: string;
  cellPhone: string;
  city: string;
  cqas: boolean;
  dob: string;
  homePhone: '';
  language: null;
  smsOptIn: false;
  state: '';
  streetAddress: '';
  zipCode: '';
}